#7: Add change handler class to handle changes
Add a new change handler class to process the before/after change event
hooks and let the highlighter focus on just highlighting. The highlighter
class has a single method to highlight a range - all the logic around what
ranges should be highlighted or not is left to the change handler.