#7: Setup basic structure of change handling with memory
Add new classes and interfaces for storing data in memory to determine
whether or not to highlight/log a change. The classes allow storing a
sheet and a range in a memory object and checking if the memory matches
a given sheet/range. The purpose is to allow setting the memory in the
`BeforeChange` method and check it against the sheet/range provided in the
`AfterChange` method.
Add placeholder tests.