#10: Add new property on the memory object for max range size
Add a new property on the change handler memory object to allow setting
the max range size that gets loaded into memory when the memory is set.
The value defaults to 15,000 cells.
Add a new test to ensure the new functionality works as expected. If no
range data is loaded into memory the comparison will return the data is
different even when called with the same sheet and range objects.