#12: Update example project to use application events
Replace the code in the example project hooking the change logger into
workbook level events to instead hook it into the application level events.
For some reason the application level events are much more reliable.
Catch exceptions when reading data from the range and log an error.
Tweak the change handler so that it does not set memory after a change.
The purpose was to set the memory to the cell that was changed if that cell
does not match the one already in memory. But that is the responsibility
of the client code, the library should not make that kind of assumption.