#13: Update memory comparison object to include before and after data
Update the `IMemoryComparison` interface and class to include the data
before and after the change. This allows the highlighter more flexibility
when processing highlighting.
Refactor the change handler code that determines if an actual change was
made to split out more logic into separate methods.
Promote another nullable type warning to an error: CS8604.