#15: Remove static change logger manager
Replace the static change logger manager and instead use the API class to
store all state. This is to avoid hardcoded references to static classes
to facilitate testing and modularity.
Update the change handler to allow injecting a custom highlighter.
Update the simple change highlighter to take the highlight colour as a
construction parameter instead of referencing a static configuration
object.
Add a static log manager to keep a static reference of the `ILogger`
instance being used.