#8: Fix check for row/column operations
Update the code checking for row/column add/removes to correct the logical
statements.
Remove condition on the row/column size of the newly changed range being
1 to allow detecting multiple rows/columns being added/removed at once.
Update the simple highlighter class to prevent highlighting row/column
deletes. These should not be highlighted because it will cause the row
below/column to the right to be highlighted. This makes no sense after a
delete operation.