#16: Support for setting multiple handlers
Refactor the change processor to support multiple handlers. When a change
is processed all handlers are called. Handlers are defined in a set, so
no ordering is guaranteed.
Rename classes to make sense with the new code structure.
Add method on the API to remove all handlers or to add a default set of
handlers.