#26: Move caching code to change processor
To ensure the handlers are called with the exact sheet and range objects
that were passed in to the change handler library, the sheet and range
data is now cached at the change processor level instead of at the API
level.
Add interfaces for the new cached sheet and range objects and rename the
delegate variables to `Raw...` instead of `_delegate`.