#27: Replace constructor assignments with lazy instanciation
Update the change handler API and change processor classes to use lazy
initialisation for properties that should never be null instead of
creating them in the constructor. This makes them safer by avoiding
issues if new constructors are added.
Remove test that is not adding any value and causing confusion.