ci: simplify the gitlab pipeline
Update the Gitlab CI pipeline to use the makefile to update the package
version and build the library instead of calling the scripts directly.
Replace the build and publish steps with a single `package` step that
builds the library on each commit. Unlike the previous `publish` step,
this runs on every branch. This simplifies the logic for the pipeline and
helps ensure that a successful build on a pull request will translate to a
successful build on the `main` branch.