feat: rename library to remove csharp reference (#35)
Rename the main and test projects from `CsharpExcelChangeHandler` to
`ExcelChangeHandler` to avoid the reference to `Csharp`. The specific
language the library was written in is irrelevant and should be removed to
avoid cluttering the name.
This change applies to the library itself as well as to the namespace of
all classes the library contains.

BREAKING CHANGE: Library and all namespaces renamed to `ExcelChangeHandler`