feat: include size of range in change properties
Update the `IChangeProperties` interface to include the number of cells in
the modified range. This ensures that it is possible to log the number of
changed cells even if the range data has not been loaded (e.g. because the
range size exceeds the memory limit).
Add new unit tests to ensure the new field behaves as expected.