feat: support system.action as change handler (#29)
Add a new API method to allow adding a System.Action as a change handler.
In the background this System.Action is wrapped in a class which implements
the `IChangeHandler` interface, allowing all downstream functionality to
continue to work in the same way.
Add unit test to ensure any added System.Action handlers get correctly
invoked when a change is detected.