feat: support for change event filters (#21)
Add support for defining change event filters to allow certain changes to
be skipped entirely. This can be used to skip changes on certain sheets or
ranges, or skip changes that meet certain criteria (e.g. large number of
cells).
The code structure for change event filters is very similar to the change
handler logic.
Add unit tests to ensure the filtering works as expected.
Update README with content on filtering.